# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LaraBook API",
        "description": "LaraBook API",
        "contact": {
            "email": "hubertlipinskipl@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/",
            "description": "Localhost server"
        }
    ],
    "paths": {
...
        "/book": {
            "get": {
                "tags": [
                    "book"
                ],
                "summary": "Get user's book list",
                "description": "Get user's book list",
                "operationId": "getList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Pagination page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Pagination page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/BookCollection"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "Password Based": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "book"
                ],
                "summary": "Store book",
                "description": "Store book",
                "operationId": "store",
                "requestBody": {
                    "$ref": "#/components/requestBodies/BookStoreRequest"
                },
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "properties": {
                                                "book": {
                                                    "$ref": "#/components/schemas/BookResource"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "Password Based": []
                    }
                ]
            }
        },
        "/book/{book_id}": {
            "put": {
                "tags": [
                    "book"
                ],
                "summary": "Update book",
                "description": "Update book",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "book_id",
                        "in": "path",
                        "description": "Book id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/BookUpdateRequest"
                },
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "properties": {
                                                "book": {
                                                    "$ref": "#/components/schemas/BookResource"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "Password Based": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "book"
                ],
                "summary": "Delete book",
                "description": "Delete book",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "book_id",
                        "in": "path",
                        "description": "Book id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "example": "Book deleted"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "Password Based": []
                    }
                ]
            }
        },
...
        "/user/{user_id}": {
            "put": {
                "tags": [
                    "user"
                ],
                "summary": "Update user",
                "description": "Update user",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UserUpdateRequest"
                },
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "$ref": "#/components/schemas/UserResource"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "Password Based": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "user"
                ],
                "summary": "Delete user",
                "description": "Delete user",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "example": "User deleted"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "Password Based": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "BookCollection": {
                "title": "BookCollection",
                "description": "Book collection",
                "properties": {
                    "books": {
                        "title": "books",
                        "description": "Array of books",
                        "type": "array",
                        "format": "array",
                        "items": {
                            "$ref": "#/components/schemas/BookResource"
                        }
                    },
                    "pagination": {
                        "$ref": "#/components/schemas/PaginationResource"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "BookCollection"
                }
            },
            "BookResource": {
                "title": "BookResource",
                "description": "Book resource",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Book id",
                        "format": "integer",
                        "example": 1
                    },
                    "title": {
                        "title": "title",
                        "description": "Book title",
                        "format": "string",
                        "example": "Title"
                    },
                    "description": {
                        "title": "description",
                        "description": "Book description",
                        "format": "string",
                        "example": "Description"
                    },
                    "short_description": {
                        "title": "short_description",
                        "description": "Book short description",
                        "format": "string",
                        "example": "Short description"
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "User created at",
                        "format": "string",
                        "example": "2021-09-06T20:58:20.000000Z"
                    },
                    "updated_at": {
                        "title": "created_at",
                        "description": "User updated at",
                        "format": "string",
                        "example": "2021-09-06T20:58:20.000000Z"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "BookResource"
                }
            },
            "PaginationResource": {
                "title": "PaginationResource",
                "description": "Pagination resource",
                "properties": {
                    "total": {
                        "title": "total",
                        "description": "Total number of items in the data store",
                        "format": "int64",
                        "example": "0"
                    },
                    "count": {
                        "title": "count",
                        "description": "Number of items for the current page",
                        "format": "int64",
                        "example": 0
                    },
                    "per_page": {
                        "title": "per_page",
                        "description": "Number of items to be shown per page",
                        "format": "int64",
                        "example": 10
                    },
                    "current_page": {
                        "title": "current_page",
                        "description": "Current page number",
                        "format": "int64",
                        "example": 1
                    },
                    "total_pages": {
                        "title": "total_pages",
                        "description": "Last available page number",
                        "format": "int64",
                        "example": 0
                    },
                    "links": {
                        "properties": {
                            "previous_page": {
                                "type": "string",
                                "example": "null"
                            },
                            "next_page": {
                                "type": "string",
                                "example": "null"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PaginationResource"
                }
            },
            "UserResource": {
                "title": "UserResource",
                "description": "User resource",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "User id",
                        "format": "integer",
                        "example": 1
                    },
                    "firstname": {
                        "title": "firstname",
                        "description": "User firstname",
                        "format": "string",
                        "example": "Jan"
                    },
                    "lastname": {
                        "title": "lastname",
                        "description": "User lastname",
                        "format": "string",
                        "example": "Kowalski"
                    },
                    "email": {
                        "title": "email",
                        "description": "User email",
                        "format": "string",
                        "example": "user@email.com"
                    },
                    "slug": {
                        "title": "slug",
                        "description": "User slug",
                        "format": "string",
                        "example": "jan-kowalski"
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "User created at",
                        "format": "string",
                        "example": "2021-09-06T20:58:20.000000Z"
                    },
                    "updated_at": {
                        "title": "created_at",
                        "description": "User updated at",
                        "format": "string",
                        "example": "2021-09-06T20:58:20.000000Z"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserResource"
                }
            }
        },
...
        "requestBodies": {
            "BookStoreRequest": {
                "description": "Book store request",
                "required": true,
                "content": {
                    "multipart/form-data": {
                        "schema": {
                            "required": [
                                "title",
                                "description",
                                "short_description"
                            ],
                            "properties": {
                                "title": {
                                    "title": "title",
                                    "description": "Book title",
                                    "format": "string",
                                    "example": "Title"
                                },
                                "description": {
                                    "title": "description",
                                    "description": "Book description",
                                    "format": "string",
                                    "example": "Description"
                                },
                                "short_description": {
                                    "title": "Short description",
                                    "description": "Book short description",
                                    "format": "string",
                                    "example": "Short description"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "BookUpdateRequest": {
                "description": "Book update request",
                "required": true,
                "content": {
                    "multipart/form-data": {
                        "schema": {
                            "properties": {
                                "title": {
                                    "title": "title",
                                    "description": "Book title",
                                    "format": "string",
                                    "example": "Title"
                                },
                                "description": {
                                    "title": "description",
                                    "description": "Book description",
                                    "format": "string",
                                    "example": "Description"
                                },
                                "short_description": {
                                    "title": "Short description",
                                    "description": "Book short description",
                                    "format": "string",
                                    "example": "Short description"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "UserUpdateRequest": {
                "description": "User update request",
                "required": true,
                "content": {
                    "multipart/form-data": {
                        "schema": {
                            "properties": {
                                "firstname": {
                                    "title": "firstname",
                                    "description": "User first name",
                                    "format": "string",
                                    "example": "Jan"
                                },
                                "lastname": {
                                    "title": "lastname",
                                    "description": "User last name",
                                    "format": "string",
                                    "example": "Kowalski"
                                },
                                "email": {
                                    "title": "User name",
                                    "description": "User email",
                                    "format": "string",
                                    "example": "new@email.com"
                                },
                                "slug": {
                                    "title": "slug",
                                    "description": "Slug",
                                    "format": "string",
                                    "example": "jan-kowalski"
                                },
                                "phone": {
                                    "title": "phone",
                                    "description": "User phone number",
                                    "format": "string",
                                    "example": "123456789"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "Password Based": {
                "type": "oauth2",
                "description": "Authenticate to API",
                "in": "header",
                "flows": {
                    "password": {
                        "tokenUrl": "/oauth/token",
                        "scopes": {}
                    }
                },
                "scheme": "token"
            }
        }
    },
    "security": [
        []
    ]
}